third_party/blink/web_tests/external/wpt/scroll-animations/view-timelines/block-view-timeline-current-time-vertical-rl.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] View timeline with container having vertical-rl layout
  assert_equals: Opacity with fill none at effect end time expected "0.7" but got "1"
Harness: the test ran to completion.
